reading file path variables from .env
