+++ config/alias.json
@@ -5,7 +5,7 @@
     },
     {
         "alias":"pl",
-        "command":"pull"
+        "command":"pl"
     },
     {
         "alias":"psh",
@@ -14,5 +14,9 @@
     {
         "alias":"s",
         "command":"status"
+    },
+    {
+        "alias":"b",
+        "command":"branch"
     }
 ]
