+++ src/.env
@@ -1,0 +1,2 @@
+ALIAS_FILE_PATH=/Users/paras/Documents/learning/go-git-it/config/alias.json
+VSC_FOLDER_PATH=/Users/paras/Documents/learning/tmp-vcs
