+++ src/.env-example
@@ -1,0 +1,2 @@
+ALIAS_FILE_PATH=/absolute/path/to/alias.json
+VSC_FOLDER_PATH=/absolute/path/to/dummy-vcs-folder
