+++ src/commit.go
@@ -1,0 +1,9 @@
+package main
+
+import (
+	"fmt"
+)
+
+func countCommit() {
+	fmt.Println("cool, will export your commit to github :)")
+}
