+++ src/main.go
@@ -3,6 +3,7 @@ package main
 import (
 	"fmt"
 	"os"
+
 )
 
 func main() {
@@ -18,7 +19,8 @@ func main() {
 		fmt.Printf(TerminalColors["printColor"], "Please enter command after program name. \nE.g., "+os.Args[0]+" command_name\n")
 		return
 	}
-	router(getAlias(), commandLineArgs...)
+	aliasFilePath := readEnvVariables()["ALIAS_FILE_PATH"].(string)
+	router(getAlias(aliasFilePath), commandLineArgs...)
 }
 
 func router(CommandsAlias map[string]string, commands ...string) {
@@ -41,6 +43,8 @@ func router(CommandsAlias map[string]string, commands ...string) {
 			return
 		}
 		execCmd("Couldn't execute git pull", commands...)
+	case "commit":
+		countCommit()
 	default:
 		execCmd("Couldn't find what you're searching for :( \n", commands...)
 	}
