--- src/utilEnums.go
@@ -1,7 +1,0 @@
-package main
-
-// TerminalColors is used for coloring terminal outputs :D
-var TerminalColors = map[string]string{
-	"printColor": "\033[0;36m%s\033[0m", //blue
-	"errorColor": "\033[1;31m%s\033[0m", //red
-}
